fix(db): move oauth-token-encryption helper into @stwd/db
The CLI script at scripts/encrypt-oauth-account-tokens.ts imported
@stwd/db, which doesn't resolve when the script is reached via a relative
import from packages/db's test (CI `bun test packages/db` fails to
resolve @stwd/db from a sibling not in workspace).

Move the actual encryption logic into packages/db/src/oauth-token-encryption.ts,
re-export it from @stwd/db, and have both the script and the test consume
the same module. Adds @stwd/vault to packages/db devDependencies for the test.
