+++ scripts/encrypt-oauth-account-tokens.ts
@@ -9,67 +9,18 @@
  * tokens are dropped. This script encrypts any renamed plaintext values that do
  * not yet have AES-GCM metadata. It is idempotent: rows with iv/tag/salt are
  * skipped.
+ *
+ * Implementation lives in `@stwd/db/src/oauth-token-encryption.ts`; this file
+ * is just the operator entrypoint.
  */
 
-import { accounts, createDb } from "@stwd/db";
-import { KeyStore } from "@stwd/vault";
-import { eq } from "drizzle-orm";
+import { createDb, encryptOAuthAccountPlaintextTokens } from "@stwd/db";
 
-export async function encryptOAuthAccountPlaintextTokens(
-  db: Pick<ReturnType<typeof createDb>["db"], "select" | "update">,
-  masterPassword: string,
-): Promise<number> {
-  if (!masterPassword) {
-    throw new Error("STEWARD_MASTER_PASSWORD is required to encrypt OAuth provider tokens");
-  }
-
-  const keyStore = new KeyStore(masterPassword);
-  const rows = await db.select().from(accounts);
-  let encryptedCount = 0;
-
-  for (const row of rows) {
-    const updates: Partial<typeof accounts.$inferInsert> = {};
-
-    if (
-      row.accessTokenEncrypted &&
-      !(row.accessTokenIv && row.accessTokenTag && row.accessTokenSalt)
-    ) {
-      const encrypted = keyStore.encrypt(row.accessTokenEncrypted);
-      updates.accessTokenEncrypted = encrypted.ciphertext;
-      updates.accessTokenIv = encrypted.iv;
-      updates.accessTokenTag = encrypted.tag;
-      updates.accessTokenSalt = encrypted.salt;
-    }
-
-    if (
-      row.refreshTokenEncrypted &&
-      !(row.refreshTokenIv && row.refreshTokenTag && row.refreshTokenSalt)
-    ) {
-      const encrypted = keyStore.encrypt(row.refreshTokenEncrypted);
-      updates.refreshTokenEncrypted = encrypted.ciphertext;
-      updates.refreshTokenIv = encrypted.iv;
-      updates.refreshTokenTag = encrypted.tag;
-      updates.refreshTokenSalt = encrypted.salt;
-    }
-
-    if (Object.keys(updates).length > 0) {
-      await db.update(accounts).set(updates).where(eq(accounts.id, row.id));
-      encryptedCount += 1;
-    }
-  }
-
-  return encryptedCount;
-}
-
-const isEntrypoint = process.argv[1] === new URL(import.meta.url).pathname;
-
-if (isEntrypoint) {
-  const masterPassword = process.env.STEWARD_MASTER_PASSWORD;
-  const { client, db } = createDb();
-  try {
-    const count = await encryptOAuthAccountPlaintextTokens(db, masterPassword ?? "");
-    console.log(`[oauth-token-encryption] encrypted ${count} account row(s)`);
-  } finally {
-    await client.end();
-  }
+const masterPassword = process.env.STEWARD_MASTER_PASSWORD;
+const { client, db } = createDb();
+try {
+  const count = await encryptOAuthAccountPlaintextTokens(db, masterPassword ?? "");
+  console.log(`[oauth-token-encryption] encrypted ${count} account row(s)`);
+} finally {
+  await client.end();
 }
